+++ packages/frontend/editor-ui/src/composables/useResolvedExpression.ts
@@ -1,16 +1,23 @@
 import { useNDVStore } from '@/stores/ndv.store';
 import { useWorkflowsStore } from '@/stores/workflows.store';
-import {
-	getResolvedExpression,
-	isExpression as isExpressionUtil,
-	stringifyExpressionResult,
-} from '@/utils/expressions';
+import { isExpression as isExpressionUtil, stringifyExpressionResult } from '@/utils/expressions';
 
 import { debounce } from 'lodash-es';
 import { createResultError, createResultOk, type IDataObject, type Result } from 'n8n-workflow';
 import { computed, onMounted, ref, toRef, toValue, watch, type MaybeRefOrGetter } from 'vue';
 import { useRouter } from 'vue-router';
 import { useWorkflowHelpers, type ResolveParameterOptions } from './useWorkflowHelpers';
+
+export const getResolvedExpression = (
+	expressionValue: boolean,
+	resolvedExpressionString: string,
+) => {
+	if (expressionValue && resolvedExpressionString) {
+		return resolvedExpressionString;
+	}
+
+	return '';
+};
 
 export function useResolvedExpression({
 	expression,
@@ -79,12 +86,10 @@ export function useResolvedExpression({
 
 	const debouncedUpdateExpression = debounce(updateExpression, 200);
 
-	function updateExpression(exp, str) {
+	function updateExpression() {
 		const resolved = resolve();
 		resolvedExpression.value = resolved.ok ? resolved.result : null;
 		resolvedExpressionString.value = stringifyExpressionResult(resolved, hasRunData.value);
-
-		resolvedExpressionString.value = exp && str ? resolvedExpressionString.value : '';
 	}
 
 	watch(
@@ -97,11 +102,10 @@ export function useResolvedExpression({
 		debouncedUpdateExpression,
 	);
 
-	onMounted(() => {
-		updateExpression(isExpression.value, resolvedExpressionString.value);
-	});
+	onMounted(updateExpression);
 
-	// resolvedExpressionString.value = isExpression.value && resolvedExpressionString.value ? resolvedExpressionString.value : '';
+	resolvedExpressionString.value =
+		isExpression.value && resolvedExpressionString.value ? resolvedExpressionString.value : '';
 
 	return { resolvedExpression, resolvedExpressionString, isExpression };
 }
