+++ packages/frontend/editor-ui/src/utils/expressions.test.ts
@@ -5,7 +5,6 @@ import {
 	removeExpressionPrefix,
 	stringifyExpressionResult,
 	unwrapExpression,
-	getResolvedExpression,
 } from './expressions';
 import { executionRetryMessage } from './executionUtils';
 
@@ -156,35 +155,6 @@ describe('Utils: Expressions', () => {
 
 		it('returns undefined for an unknown status', () => {
 			expect(executionRetryMessage('unknown')).toBeUndefined();
-		});
-	});
-
-	describe('getResolvedExpression', () => {
-		it('should return an expression if a resolved expression string is provided', () => {
-			const expressionValue = true;
-			const resolvedExpressionString = 'resolved expression';
-
-			const result = getResolvedExpression(expressionValue, resolvedExpressionString);
-
-			expect(result).toEqual(resolvedExpressionString);
-		});
-
-		it('should return an empty string if no resolved expression string is provided', () => {
-			const expressionValue = true;
-			const resolvedExpressionString = '';
-
-			const result = getResolvedExpression(expressionValue, resolvedExpressionString);
-
-			expect(result).toEqual('');
-		});
-
-		it('should return null if expression value is false', () => {
-			const expressionValue = false;
-			const resolvedExpressionString = 'resolved expression';
-
-			const result = getResolvedExpression(expressionValue, resolvedExpressionString);
-
-			expect(result).toEqual('');
 		});
 	});
 });
