+++ packages/frontend/editor-ui/src/utils/expressions.ts
@@ -159,14 +159,3 @@ export const shouldConvertToExpression = <T>(value: T, isSpecializedEditor = fal
 		value.includes('}}')
 	);
 };
-
-export const getResolvedExpression = (
-	expressionValue: boolean,
-	resolvedExpressionString: string,
-) => {
-	if (expressionValue && resolvedExpressionString) {
-		return resolvedExpressionString;
-	}
-
-	return '';
-};
